refactor(vite-dev-server): inline URL builders into devServer.ts
The warmupTargets.ts file was 50 lines of which ~30 were jsdoc/blank,
holding two small helpers used only by devServer.ts. Inline them as
exported functions in devServer.ts and rename the spec file
accordingly. No behavior change.
